Dashboard and account polish
