+++ frontend/app/components/NavbarComponent.vue
@@ -124,12 +124,6 @@ function onCalendarClick(event: Event) {
             </nuxt-link-locale>
           </li>
           <li>
-            <nuxt-link-locale @click="blur" to="/dashboard/shopping-list">
-              <i class="fa-solid fa-cart-shopping"/>
-              <span>{{$t('component.navbar.user.shopping_list')}}</span>
-            </nuxt-link-locale>
-          </li>
-          <li>
             <span @click="logout">
               <i class="fa-solid fa-arrow-right-from-bracket"/>
               <span>{{$t('component.navbar.user.logout')}}</span>
