+++ frontend/i18n/locales/ar.json
@@ -75,6 +75,8 @@
   "route./dashboard.desc": "اعرض خطط وجباتك، اكتشف وصفات جديدة، وأدر إعدادات حسابك.",
   "route./dashboard/account": "الحساب",
   "route./dashboard/account.desc": "اعرض وعدّل معلومات حسابك، بما في ذلك اسم المستخدم والبريد الإلكتروني وكلمة المرور.",
+  "route./dashboard/account/edit": "تعديل الملف الشخصي",
+  "route./dashboard/account/edit.desc": "إدارة إعدادات ملفك الشخصي، بما في ذلك اسم المستخدم وكلمة المرور.",
   "route./dashboard/schedule": "الجدول",
   "route./dashboard/schedule.desc": "اعرض وأدر خطط وجباتك للأسبوع. أضف وجبات جديدة وعدّل الموجودة ونظّم جدولك.",
   "route./dashboard/shopping-list": "قائمة التسوق",
