+++ frontend/i18n/locales/brl.json
@@ -75,6 +75,8 @@
   "route./dashboard.desc": "⠠⠧⠊⠑⠺⠀⠽⠕⠥⠗⠀⠍⠑⠁⠇⠀⠏⠇⠁⠝⠎,⠀⠙⠊⠎⠉⠕⠧⠑⠗⠀⠝⠑⠺⠀⠗⠑⠉⠊⠏⠑⠎,⠀⠁⠝⠙⠀⠍⠁⠝⠁⠛⠑⠀⠽⠕⠥⠗⠀⠁⠉⠉⠕⠥⠝⠞⠀⠎⠑⠞⠞⠊⠝⠛⠎.",
   "route./dashboard/account": "⠠⠁⠉⠉⠕⠥⠝⠞",
   "route./dashboard/account.desc": "⠠⠧⠊⠑⠺⠀⠁⠝⠙⠀⠑⠙⠊⠞⠀⠽⠕⠥⠗⠀⠁⠉⠉⠕⠥⠝⠞⠀⠊⠝⠋⠕⠗⠍⠁⠞⠊⠕⠝,⠀⠊⠝⠉⠇⠥⠙⠊⠝⠛⠀⠽⠕⠥⠗⠀⠥⠎⠑⠗⠝⠁⠍⠑,⠀⠑⠍⠁⠊⠇,⠀⠁⠝⠙⠀⠏⠁⠎⠎⠺⠕⠗⠙.",
+  "route./dashboard/account/edit": "⠠⠑⠙⠊⠞⠀⠠⠏⠗⠕⠋⠊⠇⠑",
+  "route./dashboard/account/edit.desc": "⠠⠍⠁⠝⠁⠛⠑⠀⠽⠕⠥⠗⠀⠏⠗⠕⠋⠊⠇⠑⠀⠎⠑⠞⠞⠊⠝⠛⠎,⠀⠊⠝⠉⠇⠥⠙⠊⠝⠛⠀⠥⠎⠑⠗⠝⠁⠍⠑⠀⠁⠝⠙⠀⠏⠁⠎⠎⠺⠕⠗⠙.",
   "route./dashboard/schedule": "⠠⠎⠉⠓⠑⠙⠥⠇⠑",
   "route./dashboard/schedule.desc": "⠠⠧⠊⠑⠺⠀⠁⠝⠙⠀⠍⠁⠝⠁⠛⠑⠀⠽⠕⠥⠗⠀⠍⠑⠁⠇⠀⠏⠇⠁⠝⠎⠀⠋⠕⠗⠀⠞⠓⠑⠀⠺⠑⠑⠅.⠀⠠⠁⠙⠙⠀⠝⠑⠺⠀⠍⠑⠁⠇⠎,⠀⠑⠙⠊⠞⠀⠑⠭⠊⠎⠞⠊⠝⠛⠀⠕⠝⠑⠎,⠀⠁⠝⠙⠀⠕⠗⠛⠁⠝⠊⠵⠑⠀⠽⠕⠥⠗⠀⠎⠉⠓⠑⠙⠥⠇⠑.",
   "route./dashboard/shopping-list": "⠠⠎⠓⠕⠏⠏⠊⠝⠛⠀⠠⠇⠊⠎⠞",
