+++ frontend/i18n/locales/zh.json
@@ -75,6 +75,8 @@
   "route./dashboard.desc": "查看您的餐食计划，发现新食谱，管理您的账户设置。",
   "route./dashboard/account": "账户",
   "route./dashboard/account.desc": "查看和编辑您的账户信息，包括用户名、电子邮件和密码。",
+  "route./dashboard/account/edit": "编辑资料",
+  "route./dashboard/account/edit.desc": "管理您的个人资料设置，包括用户名和密码。",
   "route./dashboard/schedule": "日程",
   "route./dashboard/schedule.desc": "查看和管理本周的餐食计划。",
   "route./dashboard/shopping-list": "购物清单",
